requirements.txt
beautifulsoup4==4.11.1
Django==4.1.4
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-tinymce==3.5.0
djangorestframework==3.14.0
virtualenv==20.23.0
scikit-learn==1.2.1
requests==2.28.2
Pillow==9.5.0
pandas==1.5.2
numpy==1.24.1
openai==0.28.0
mysql-connector==2.2.9
mysqlclient==2.1.1